Now Bagging (and RandomForest) can process datasets with more than 22 million instances. This one was very mysterious at first.
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka@11467 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ branches/stable-3-6/weka/src/main/java/weka/classifiers/meta/Bagging.java
@@ -497,7 +497,7 @@ public class Bagging extends RandomizableIteratedSingleClassifierEnhancer
           + "out-of-bag error is to be calculated!");
     }
 
-    int bagSize = data.numInstances() * m_BagSizePercent / 100;
+    int bagSize = (int) (data.numInstances() * (m_BagSizePercent / 100.0));
     Random random = new Random(m_Seed);
 
     boolean[][] inBag = null;
